minio service update and video/file related fixes

+++ Backend/Repositories/Abstraction/BaseAsyncCrudRepository.cs
@@ -22,7 +22,7 @@ public class BaseAsyncCrudRepository<T>(DbContext context, DbSet<T> set) : IAsyn
         await SaveAsync();
     }
 
-    public async Task<bool> DeleteById(long id)
+    public async Task<bool> DeleteById(int id)
     {
         var entity = await set.FindAsync(id);
         return await Delete(entity);
